refactor: review module syntax
