+++ templates/vsc/js/default-bot/agent.js
