+++ templates/vsc/ts/basic-custom-engine-agent/package.json.tpl
@@ -5,9 +5,6 @@
         "teamsAppId": null
     },
     "description": "Basic Custome Engine Agent with Microsoft 365 Agents SDK",
-    "engines": {
-        "node": "18 || 20 || 22"
-    },
     "author": "Microsoft",
     "license": "MIT",
     "main": "./lib/src/index.js",
