+++ templates/vsc/ts/default-bot/package.json.tpl
@@ -2,9 +2,6 @@
     "name": "{{SafeProjectNameLowerCase}}",
     "version": "1.0.0",
     "description": "Microsoft 365 Agents Toolkit echo bot sample",
-    "engines": {
-        "node": "18 || 20 || 22"
-    },
     "author": "Microsoft",
     "license": "MIT",
     "main": "./lib/index.js",
